fix incomplete database on new scan

+++ src/db/database-manager.ts
@@ -434,23 +434,29 @@ export class DatabaseManager {
 			await insertTracks();
 		}
 
-		await this.db.delete(Schema.songs).where(ne(Schema.songs.syncId, syncId));
-		await this.db.delete(Schema.albums).where(ne(Schema.albums.syncId, syncId));
-		await this.db
-			.delete(Schema.artists)
-			.where(ne(Schema.artists.syncId, syncId));
-		await this.db
-			.delete(Schema.albumArtists)
-			.where(ne(Schema.albumArtists.syncId, syncId));
-		await this.db
-			.delete(Schema.songArtists)
-			.where(ne(Schema.songArtists.syncId, syncId));
-		await this.db.delete(Schema.genres).where(ne(Schema.genres.syncId, syncId));
-		await this.db
-			.delete(Schema.albumGenres)
-			.where(ne(Schema.albumGenres.syncId, syncId));
-		await this.db
-			.delete(Schema.songGenres)
-			.where(ne(Schema.songGenres.syncId, syncId));
+		if (!onlyNew) {
+			await this.db.delete(Schema.songs).where(ne(Schema.songs.syncId, syncId));
+			await this.db
+				.delete(Schema.albums)
+				.where(ne(Schema.albums.syncId, syncId));
+			await this.db
+				.delete(Schema.artists)
+				.where(ne(Schema.artists.syncId, syncId));
+			await this.db
+				.delete(Schema.albumArtists)
+				.where(ne(Schema.albumArtists.syncId, syncId));
+			await this.db
+				.delete(Schema.songArtists)
+				.where(ne(Schema.songArtists.syncId, syncId));
+			await this.db
+				.delete(Schema.genres)
+				.where(ne(Schema.genres.syncId, syncId));
+			await this.db
+				.delete(Schema.albumGenres)
+				.where(ne(Schema.albumGenres.syncId, syncId));
+			await this.db
+				.delete(Schema.songGenres)
+				.where(ne(Schema.songGenres.syncId, syncId));
+		}
 	}
 }
